+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/sdkconfiguration.py
@@ -2,9 +2,10 @@
 
 import requests
 from dataclasses import dataclass
-from typing import Dict, Tuple
+from typing import Dict, Tuple, Callable, Union
 from .utils.retries import RetryConfig
 from .utils import utils
+from codatsyncexpenses.models import shared
 
 
 SERVERS = [
@@ -15,14 +16,14 @@ SERVERS = [
 @dataclass
 class SDKConfiguration:
     client: requests.Session
-    security_client: requests.Session
+    security: Union[shared.Security,Callable[[], shared.Security]] = None
     server_url: str = ''
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = 'prealpha'
-    sdk_version: str = '0.2.0'
-    gen_version: str = '2.159.2'
-    user_agent: str = 'speakeasy-sdk/python 0.2.0 2.159.2 prealpha codat-sync-for-expenses-version-1'
+    sdk_version: str = '0.3.0'
+    gen_version: str = '2.209.0'
+    user_agent: str = 'speakeasy-sdk/python 0.3.0 2.209.0 prealpha codat-sync-for-expenses-version-1'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: