+++ previous-versions/sync-for-expenses-version-1/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.2.0] previous-versions/sync-for-expenses-version-1
 ### Releases
 - [PyPI v0.2.0] https://pypi.org/project/codat-sync-for-expenses-version-1/0.2.0 - previous-versions/sync-for-expenses-version-1
+
+## 2023-12-04 01:25:27
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses-v1.yaml
+- Speakeasy CLI 1.125.0 (2.209.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.3.0] previous-versions/sync-for-expenses-version-1
+### Releases
+- [PyPI v0.3.0] https://pypi.org/project/codat-sync-for-expenses-version-1/0.3.0 - previous-versions/sync-for-expenses-version-1